requirements.txt
oaas-sdk-py 
grpcio
grpcio-tools
oaas-sdk-grpc
aiohttp~=3.9.5
protobuf